Rename components to PascalCase
